Always enable syntax highlighting, minor refactor
- Extract styles into module.css
- Remove obsolete types
