+++ src/components/Generator/GeneredTaskfile/GeneratedTaskfile.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ReactElement, useState } from 'react';
+import { ReactElement } from 'react';
 import { taskfile } from '@/components/Generator/GeneredTaskfile/taskfile';
 import { useFormContext } from 'react-hook-form';
 import { GeneratorSettings } from '@/components/Generator/Generator';
@@ -8,8 +8,6 @@ import CopyToClipboard from '@/components/Generator/GeneredTaskfile/Copy';
 import { highlighter } from './Highlighter';
 
 const GeneratedTaskfile = (): ReactElement => {
-	const [showHighlighting, setShowHighlighting] = useState(true);
-
 	const form = useFormContext<GeneratorSettings>();
 
 	const settings = form.watch();
@@ -19,9 +17,7 @@ const GeneratedTaskfile = (): ReactElement => {
 	return (
 		<>
 			<CopyToClipboard onCopy={() => navigator.clipboard.writeText(resultTaskfile)} />
-			<pre onClick={() => setShowHighlighting(!showHighlighting)}>
-				{showHighlighting ? highlighter(resultTaskfile) : resultTaskfile}
-			</pre>
+			<pre>{highlighter(resultTaskfile)}</pre>
 		</>
 	);
 };
