--- src/components/Generator/GeneredTaskfile/Highlighter/styles.ts
@@ -1,18 +1,0 @@
-import { StylesConfig } from './types';
-
-export const styles: StylesConfig = {
-	line: {
-		display: 'flex',
-	},
-	colors: {
-		white: { color: '#fff' },
-		purple: { color: '#a855f7' },
-		yellow: { color: '#fde047' },
-		green: { color: '#22c55e' },
-		blue: { color: '#60a5fa' },
-		pink: { color: '#ec4899' },
-		lightYellow: { color: '#fef08a' },
-		lightBlue: { color: '#93c5fd' },
-		gray: { color: '#888589' },
-	},
-};
