+++ src/components/Generator/GeneredTaskfile/Highlighter/types.ts
@@ -1,12 +1,5 @@
 import { ReactElement } from 'react';
 
-type ColorName = 'white' | 'purple' | 'yellow' | 'green' | 'blue' | 'pink' | 'lightYellow' | 'lightBlue' | 'gray';
-
-export type StylesConfig = {
-	line: { display: string };
-	colors: Record<ColorName, { color: string }>;
-};
-
 export enum RendererType {
 	EmptyLines,
 	FunctionDefinitions,
